requirements.txt
contextlib2==0.5.4
coreapi==1.32.0
coverage==4.2
Django==1.10
django-environ==0.4.0
django-rest-swagger==0.3.8
djangorestframework==3.4.1
itypes==1.1.0
openapi-codec==1.0.0
psycopg2==2.6.1
py==1.4.31
pymongo==3.3.0
pytest==2.9.2
pytest-cov==2.3.0
pytest-django==2.9.1
python-dateutil==2.5.3
PyYAML==3.11
raven==5.21.0
requests==2.10.0
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
